spanish common voice: AM_learning 20 epochs, weight of 0.01 for LM in rcognition. meaningful results

+++ espnet/dataset_utils/transformers/spanish_common_voice.py
@@ -73,10 +73,15 @@ class CommonVoiceKaldiTransformer(AbstractDataTransformer):
         new_file_name = source_path.split("/")[-1][:-4] + '.wav'
         upsampled_wav_path = Path(kaldi_audio_files_dir, new_file_name + '.upsample')
         downsampled_wav_path = Path(kaldi_audio_files_dir, new_file_name)
-        sound = AudioSegment.from_mp3(source_path)
-        sound.export(upsampled_wav_path, format="wav")
-        os.system("sox -v 0.80 '%s' -r 16000 -b 16 -c 1 %s" % (upsampled_wav_path, downsampled_wav_path))
-        os.remove(upsampled_wav_path)
+        if downsampled_wav_path.exists():
+            return
+        try: 
+            sound = AudioSegment.from_mp3(source_path)
+            sound.export(upsampled_wav_path, format="wav")
+            os.system("sox -v 0.80 '%s' -r 16000 -b 16 -c 1 %s" % (upsampled_wav_path, downsampled_wav_path))
+            os.remove(upsampled_wav_path)
+        except:
+            print(f"convert_to_wav_from_mp3: An exception occurred, source_path({source_path})")
 
     def generate_arrays(self, data: pd.DataFrame):
